fix,test(padma/resolver): lodda things
couldn't describe them, should've commited them in smaller chunks
but anyway it works now :>

+++ padma-core/src/defs/models.rs
@@ -49,7 +49,7 @@ pub enum DispatchKind {
 #[derive(Debug, Clone, PartialEq)]
 pub enum BlockArgument {
     Dispatch(Dispatch),
-    Argument { index: u32 },
+    Argument { index: usize },
     Literal(Literal),
     This,
 }
